Don't try to use nonexistent libpng optimizations
The embedded copy of libpng doesn't contain code for processor-specific
hardware optimizations, but it may still try to use them in some cases,
resulting in compiler errors.

Ensure optimizations are disabled by defining several macros. These are
the same macros that upstream libpng defines when hardware optimizations
are disabled in the build configuration.