Don't try to use nonexistent libpng optimizations
The embedded copy of libpng doesn't contain code for processor-specific
hardware optimizations, but it may still try to use them in some cases,
resulting in linker errors.

juce_PNGLoader.cpp already sets `PNG_ARM_NEON_OPT` to 0 to disable
optimizations on ARM, but this should be done for all architectures,
especially PowerPC, where builds currently fail.

On x86, this isn't strictly necessary, because libpng optimizations are
opt-in for x86, rather than opt-out as with all other architectures, but
for completeness and robustness it is also included here.

The macros to disable optimizations on other platforms come straight
from libpng's own build files; this is how it disables optimizations
when compiled without support for them. `PNG_ARM_NEON_OPT` is already
one of those macros; this commit simply adds the other three.